implement ecs_compatibility mode
Uses a generic `Chooser`, which should probably be extracted to a separate
zero-dependency library for use by other plugins. The `Chooser` allows us to
define `ecs_compatibility` modes for `disabled` and `v1` side-by-side instead
of in separate and potentially-diverging blocks.

Fixes a missing ip address of source when a TCP Proxy is present.